refactor(map): extract map layer/draw helpers; preserve curated paths/pois
Addresses CodeRabbit review on #64:
- Move initLeafletMap, flyAndReveal, drawLandmarks, drawRemotePlaces (+ remote/
  landmark data) out of MapCanvas.tsx into a new mapLayers.ts. MapCanvas drops
  from 319 to 183 lines; both files are now under the 200-line guideline.
- fetch-remote-places.mjs now reads the existing JSON and merges hand-curated
  paths/pois back per-id before writing, so a regeneration no longer silently
  drops the arboretum's footpath network + labelled points. Warns that
  outline/area are re-fetched and manual trims must be re-applied.
